requirements.txt
absl-py==0.7.0
astor==0.7.1
blinker==1.4
bz2file==0.98
cachetools==3.0.0
cloudpickle==0.6.1
configparser==3.5.0
cycler==0.10.0
dask==1.1.0
decorator==4.3.0
gast==0.2.2
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
grpcio==1.18.0
h5py==2.9.0
httplib2==0.12.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
networkx==2.2
nibabel==2.3.3
numpy==1.14.5
oauth2client==4.1.3
packaging==19.0
pandas==0.23.4
Pillow==5.4.1
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
PyDrive==1.3.1
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
PyWavelets==1.0.1
PyYAML==3.13
rsa==4.0
scikit-image==0.14.2
scipy==1.2.0
SimpleITK==1.2.0
six==1.12.0
tensorboard==1.10.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
toolz==0.9.0
uritemplate==3.0.0
Werkzeug==0.14.1
